requirements.txt
python-telegram-bot[ext]
pytz
google-generativeai